refactor: export subpaths from pwa package (wip)
