+++ pwa/package.json
@@ -22,10 +22,20 @@
         "workbox-routing": "^6.1.5",
         "workbox-strategies": "^6.1.5"
     },
-    "main": "./build/cjs/index.js",
-    "module": "./build/es/index.js",
+    "main": "./build/cjs/browser/index.js",
+    "module": "./build/es/browser/index.js",
     "exports": {
-        "import": "./build/es/index.js",
-        "require": "./build/cjs/index.js"
+        ".": {
+            "import": "./build/es/browser/index.js",
+            "require": "./build/cjs/browser/index.js"
+        },
+        "./sw": {
+            "import": "./build/es/sw/index.js",
+            "require": "./build/cjs/sw/index.js"
+        },
+        "./browser": {
+            "import": "./build/es/browser/index.js",
+            "require": "./build/cjs/browser/index.js"
+        }
     }
 }
