+++ pwa/src/browser/index.js
@@ -1,0 +1,2 @@
+export { OfflineInterface } from './offline-interface'
+export { checkForSWUpdateAndReload } from '../lib/registration'
