+++ pwa/src/browser/offline-interface.js
