--- pwa/src/index.js
@@ -1,6 +1,0 @@
-export { setUpServiceWorker } from './service-worker/service-worker.js'
-export { OfflineInterface } from './offline-interface/offline-interface.js'
-export {
-    checkForUpdates,
-    checkForSWUpdateAndReload,
-} from './lib/registration.js'
