+++ shell/src/service-worker.js
@@ -1,3 +1,3 @@
-import { setUpServiceWorker } from '@dhis2/pwa'
+import { setUpServiceWorker } from '@dhis2/pwa/sw'
 
 setUpServiceWorker()
